refractored code and added run.sh file to start
