+++ run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -e
+PYTHONPATH=server uvicorn request:app --reload --host 0.0.0.0 --port 8000
